Moving Component to Helpers
