+++ clients/admin-ui/src/features/connector-templates/SaaSVersionModal.tsx
@@ -1,9 +1,8 @@
 import { Button, Flex, Modal, Spin, Tabs, Text } from "fidesui";
 import React from "react";
 
-import ClipboardButton from "~/features/common/ClipboardButton";
 import { MODAL_SIZE } from "~/features/common/modals/modal-sizes";
-import { Editor } from "~/features/common/yaml/helpers";
+import { EditorWithCopy } from "~/features/common/yaml/helpers";
 
 import {
   useGetConnectorTemplateVersionConfigQuery,
@@ -15,28 +14,6 @@ interface SaaSVersionContentProps {
   version: string;
 }
 
-const EditorWithCopy = ({ yaml }: { yaml: string }) => (
-  <div className="relative">
-    <Editor
-      defaultLanguage="yaml"
-      value={yaml}
-      height="60vh"
-      options={EDITOR_OPTIONS}
-      theme="light"
-    />
-    <div className="absolute right-2 top-2 z-10">
-      <ClipboardButton copyText={yaml} size="small" />
-    </div>
-  </div>
-);
-
-const EDITOR_OPTIONS = {
-  readOnly: true,
-  minimap: { enabled: false },
-  fontSize: 13,
-  fontFamily: "Menlo",
-  scrollBeyondLastLine: false,
-};
 
 const SaaSVersionContent = ({
   connectorType,
